LayoutTests/accessibility/win/text-role-expected.txt

This test verifies that, on Windows, text nodes return "editable text" for their role value.

Bug 32631 - MSAA: Accessibility role of text nodes is wrong

FAIL e.role should be editable text. Was AXRole: AXStaticText.
